server/requirements.txt
fastapi
uvicorn
scikit-learn
pydantic
python-dotenv